requirements-base.txt
pytest==9.0.2
pytest-html==4.1.1
pytest-xdist==3.8.0
python-dotenv==1.2.1